Refactored build helpers into individual files #248
