+++ tests/build_helpers/dpkg.py
@@ -1,0 +1,41 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Tests for the helper for building projects from source."""
+
+from __future__ import unicode_literals
+
+import unittest
+
+
+class DPKGBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build dpkg packages (.deb)."""
+
+  # TODO: add tests.
+
+
+class ConfigureMakeDPKGBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build dpkg packages (.deb)."""
+
+  # TODO: add tests.
+
+
+class ConfigureMakeSourceDPKGBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build source dpkg packages (.deb)."""
+
+  # TODO: add tests.
+
+
+class SetupPyDPKGBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build dpkg packages (.deb)."""
+
+  # TODO: add tests.
+
+
+class SetupPySourceDPKGBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build source dpkg packages (.deb)."""
+
+  # TODO: add tests.
+
+
+if __name__ == '__main__':
+  unittest.main()
