+++ tests/build_helpers/msi.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Tests for the helper for building projects from source."""
+
+from __future__ import unicode_literals
+
+import unittest
+
+
+class MSIBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build Microsoft Installer packages (.msi)."""
+
+  # TODO: add tests.
+
+
+class ConfigureMakeMSIBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build Microsoft Installer packages (.msi)."""
+
+  # TODO: add tests.
+
+
+class SetupPyMSIBuildHelperTest(unittest.TestCase):
+  """Tests for the helper to build Microsoft Installer packages (.msi)."""
+
+  # TODO: add tests.
+
+
+if __name__ == '__main__':
+  unittest.main()
