+++ tests/build_helpers/osc.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Tests for the helper for building projects from source."""
+
+from __future__ import unicode_literals
+
+import unittest
+
+
+class OSCBuildHelperTest(unittest.TestCase):
+  """Tests for the build with osc for the openSUSE build service."""
+
+  # TODO: add tests.
+
+
+class ConfigureMakeOSCBuildHelperTest(unittest.TestCase):
+  """Tests for the build with osc for the openSUSE build service."""
+
+  # TODO: add tests.
+
+
+class SetupPyOSCBuildHelperTest(unittest.TestCase):
+  """Tests for the build with osc for the openSUSE build service."""
+
+  # TODO: add tests.
+
+
+if __name__ == '__main__':
+  unittest.main()
