+++ tests/build_helpers/pkg.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Tests for the helper for building projects from source."""
+
+from __future__ import unicode_literals
+
+import unittest
+
+
+# TODO: add PKGBuildHelper tests.
+# TODO: add ConfigureMakePKGBuildHelper tests.
+# TODO: add SetupPyPKGBuildHelper tests.
+
+
+if __name__ == '__main__':
+  unittest.main()
